Move SpatialPooler and TemporalMemory to Algorithms package where they belong
