+++ [ htm.java Unit Test ] TemporalMemoryTest.launch
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <launchConfiguration type="org.eclipse.jdt.junit.launchconfig">
 <listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_PATHS">
-<listEntry value="/htm.java/src/test/java/org/numenta/nupic/research/TemporalMemoryTest.java"/>
+<listEntry value="/htm.java/src/test/java/org/numenta/nupic/algorithms/TemporalMemoryTest.java"/>
 </listAttribute>
 <listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_TYPES">
 <listEntry value="1"/>
@@ -25,7 +25,7 @@
 </listAttribute>
 <stringAttribute key="org.eclipse.jdt.launching.CLASSPATH_PROVIDER" value="org.eclipse.m2e.launchconfig.classpathProvider"/>
 <booleanAttribute key="org.eclipse.jdt.launching.DEFAULT_CLASSPATH" value="false"/>
-<stringAttribute key="org.eclipse.jdt.launching.MAIN_TYPE" value="org.numenta.nupic.research.TemporalMemoryTest"/>
+<stringAttribute key="org.eclipse.jdt.launching.MAIN_TYPE" value="org.numenta.nupic.algorithms.TemporalMemoryTest"/>
 <stringAttribute key="org.eclipse.jdt.launching.PROJECT_ATTR" value="htm.java"/>
 <stringAttribute key="org.eclipse.jdt.launching.SOURCE_PATH_PROVIDER" value="org.eclipse.m2e.launchconfig.sourcepathProvider"/>
 </launchConfiguration>
