+++ src/main/java/org/numenta/nupic/ComputeCycle.java
@@ -26,10 +26,10 @@ import java.util.LinkedHashSet;
 import java.util.Map;
 import java.util.Set;
 
+import org.numenta.nupic.algorithms.TemporalMemory;
 import org.numenta.nupic.model.Cell;
 import org.numenta.nupic.model.Column;
 import org.numenta.nupic.model.DistalDendrite;
-import org.numenta.nupic.research.TemporalMemory;
 
 /**
  * Contains a snapshot of the state attained during one computational
