+++ src/main/java/org/numenta/nupic/network/ManualInput.java
@@ -26,8 +26,8 @@ import java.util.Map;
 
 import org.numenta.nupic.algorithms.CLAClassifier;
 import org.numenta.nupic.algorithms.ClassifierResult;
+import org.numenta.nupic.algorithms.SpatialPooler;
 import org.numenta.nupic.encoders.Encoder;
-import org.numenta.nupic.research.SpatialPooler;
 import org.numenta.nupic.util.ArrayUtils;
 import org.numenta.nupic.util.NamedTuple;
