+++ src/main/java/org/numenta/nupic/research/README.md
@@ -1,0 +1,1 @@
+# This is a holding space for research, research ports, and experiments
