+++ src/test/java/org/numenta/nupic/ConnectionsTest.java
@@ -10,9 +10,9 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import org.junit.Test;
+import org.numenta.nupic.algorithms.TemporalMemory;
 import org.numenta.nupic.model.Cell;
 import org.numenta.nupic.model.Column;
-import org.numenta.nupic.research.TemporalMemory;
 
 
 public class ConnectionsTest {
